requirements.txt
asgiref==3.2.7
astroid==2.4.1
autopep8==1.5.2
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.7
django-cors-headers==3.2.1
django-filter==2.2.0
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
imutils==0.5.3
isort==4.3.21
joblib==0.14.1
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
numpy==1.18.4
opencv-python==4.2.0.34
pandas==1.0.5
Pillow==7.1.2
psycopg2==2.8.5
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.5.2
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
razorpay==1.2.0
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sqlparse==0.3.1
threadpoolctl==2.0.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
whitenoise==5.0.1
wrapt==1.12.1
xgboost==1.0.2
XlsxWriter==1.2.9
zipp==3.1.0
